% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{plot.ryx}
\usage{
\method{plot}{ryx}(x)
}
\arguments{
\item{x}{A ryx object}
}
\value{
A ggplot graph of the correlations
}
\description{
Plots a ryx object as a scatter plot where the x axis is correlation and y-axis are independent
variables
}
\examples{
# plot(ryx(Boston, y="medv"))
plot(ryx(mtcars, y="mpg", x=c("wt", "vs")))

}
