% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{print}
\usage{
\method{print}{ryx}(x)
}
\arguments{
\item{x}{A ryx object}
}
\value{
A dataframe contaning the correlation and significance of the
specified object. Has variable, r, p, and sigif column.
}
\description{
Takes an ryx object and prints a correlation dataframe displaying
correlation and significance for each variable
}
\examples{
\dontrun{
print(ryx(Boston, y="medv"))
print(ryx(mtcars, y="mpg", x=c("wt", "vs"))
}
}
