% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{ryx}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{The dataset that contains that relevant variables}

\item{y}{The name of a numeric variable in the data frame that will be the
independent variable}

\item{x}{is a character vector with the names of one or more numeric
variables in the data frame. Defaults to all numeric columns in the dataframe}
}
\value{
An obect of type 'ryx' that has 3 attributes, The dependent variable,
the list of independent variables, and a dataframe showing correlation and
significance of each independent variable with the dependent.
}
\description{
A function that calculates correlations between a dependent variable y and
one or more independent x variables.
}
\examples{
\dontrun{
#calculate correlation on 'medv' in the dataframe, uses all numeric columns
ryx(Boston, y="medv")
#can specify indepenent vars, and be assigned to a var
x <- ryx(mtcars, y="mpg", x=c("wt", "vs"))
}

}
