% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{summary.ryx}
\usage{
\method{summary}{ryx}(x)
}
\arguments{
\item{x}{A 'ryx' object}
}
\value{
A string
}
\description{
Provides some summary statistics about the correlations, such as:
median correlation, range of correlation, how many variables were
statistically significant.
}
\examples{

# summary(ryx(Boston, y="medv"))
summary(ryx(mtcars, y="mpg", x=c("wt", "vs")))

}
